player character somewhat playable and is under prefabs.
player can now:
attack (+with combos, "combos" right now are only the same attack, but it works with more, but need more animations for that)
use skills
change weapons
throw grenade (with mouse and controller)

problems i had: combocounter because i dont know on which enemy to test. problems with dash.